requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
Flask
Werkzeug
aiohttp
beautifulsoup4
cloudscraper
hjson
pymongo
python_dotenv
requests
scikit_learn
scrapyscript
selenium
selenium_requests
termcolor
numpy
dnspython
amazoncaptcha
coloredlogs
mypy
Flask-RESTful
